added gps window, basically
There's a GPS button on the map. Click it, and it'll open a tableview
showing an array of recent GPS check-ins (currently running constantly
in the background - I'll add a 'GPS on/off' button in my help window).
Try it out while you walk around your house!

The idea is that the array of check-ins would simply be sent to a St
Fagan's central server (I've created a UUID that's sent with it). I
guess there's no need to add this functionality in now, though.